+++ config/locales/en.yml
@@ -1216,7 +1216,7 @@ en:
             query:
               blank: Enter a URN (unique reference number) or school name
               too_short: Enter at least two characters
-        support/providers/schools/urn_form:
+        urn_form:
           attributes:
             values:
               blank: "Enter URNs"
